create pidgin str utility
