+++ src/a16_pidgin_logic/pidgin_config.py
@@ -21,82 +21,6 @@ def get_pidgin_dimens() -> set[str]:
     return set(get_pidgin_config_dict().keys())
 
 
-def pidginunit_str() -> str:
-    return "pidginunit"
-
-
-def pidgin_filename() -> str:
-    return "pidgin.json"
-
-
-def otx_bridge_str() -> str:
-    return "otx_bridge"
-
-
-def inx_bridge_str() -> str:
-    return "inx_bridge"
-
-
-def inx_label_str() -> str:
-    return "inx_label"
-
-
-def otx_label_str() -> str:
-    return "otx_label"
-
-
-def inx_name_str() -> str:
-    return "inx_name"
-
-
-def otx_name_str() -> str:
-    return "otx_name"
-
-
-def inx_tag_str() -> str:
-    return "inx_tag"
-
-
-def otx_tag_str() -> str:
-    return "otx_tag"
-
-
-def inx_road_str() -> str:
-    return "inx_road"
-
-
-def otx_road_str() -> str:
-    return "otx_road"
-
-
-def unknown_word_str() -> str:
-    return "unknown_word"
-
-
-def otx2inx_str() -> str:
-    return "otx2inx"
-
-
-def map_otx2inx_str() -> str:
-    return "map_otx2inx"
-
-
-def pidgin_name_str() -> str:
-    return "pidgin_name"
-
-
-def pidgin_label_str() -> str:
-    return "pidgin_label"
-
-
-def pidgin_tag_str() -> str:
-    return "pidgin_tag"
-
-
-def pidgin_road_str() -> str:
-    return "pidgin_road"
-
-
 def default_unknown_word() -> str:
     return "UNKNOWN"
