+++ src/a16_pidgin_logic/test__pidgin_config/test_pidgin_config_.py
@@ -7,13 +7,7 @@ from src.a08_bud_atom_logic._utils.str_a08 import (
     atom_delete,
     normal_specs_str,
 )
-from src.a16_pidgin_logic.pidgin_config import (
-    config_file_dir,
-    get_pidgin_dimens,
-    get_pidgin_config_filename,
-    get_pidgin_config_dict,
-    get_pidgin_args_dimen_mapping,
-    get_quick_pidgens_column_ref,
+from src.a16_pidgin_logic._utils.str_a16 import (
     pidginunit_str,
     pidgin_filename,
     otx_bridge_str,
@@ -33,6 +27,14 @@ from src.a16_pidgin_logic.pidgin_config import (
     pidgin_label_str,
     pidgin_tag_str,
     pidgin_road_str,
+)
+from src.a16_pidgin_logic.pidgin_config import (
+    config_file_dir,
+    get_pidgin_dimens,
+    get_pidgin_config_filename,
+    get_pidgin_config_dict,
+    get_pidgin_args_dimen_mapping,
+    get_quick_pidgens_column_ref,
     default_unknown_word,
     default_unknown_word_if_None,
 )
