+++ src/a16_pidgin_logic/test_pidgin/test_map_group.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import default_bridge_if_None
 from src.a06_bud_logic._utils.str_a06 import event_int_str, face_name_str
 from src.a16_pidgin_logic.pidgin_config import default_unknown_word_if_None
-from src.a16_pidgin_logic.pidgin_config import (
+from src.a16_pidgin_logic._utils.str_a16 import (
     otx_bridge_str,
     inx_bridge_str,
     unknown_word_str,
