+++ src/a16_pidgin_logic/test_pidgin/test_pidgin_road_.py
@@ -1,7 +1,7 @@
 from src.a01_word_logic.road import default_bridge_if_None, create_road
 from src.a06_bud_logic._utils.str_a06 import event_int_str, face_name_str
 from src.a16_pidgin_logic.pidgin_config import default_unknown_word_if_None
-from src.a16_pidgin_logic.pidgin_config import (
+from src.a16_pidgin_logic._utils.str_a16 import (
     inx_bridge_str,
     otx2inx_str,
     otx_bridge_str,
