+++ src/a17_idea_logic/test__idea_db_tool/test_pandas_tool_pidgin_files.py
@@ -2,7 +2,7 @@ from src.a00_data_toolbox.file_toolbox import save_file, create_path
 from src.a01_word_logic.road import create_road
 from src.a06_bud_logic._utils.str_a06 import acct_name_str, base_str
 from src.a16_pidgin_logic.pidgin import pidginunit_shop
-from src.a16_pidgin_logic.pidgin_config import pidgin_filename
+from src.a16_pidgin_logic._utils.str_a16 import pidgin_filename
 from src.a16_pidgin_logic._utils.env_a16 import (
     env_dir_setup_cleanup,
     get_example_face_dir,
