+++ src/a18_etl_toolbox/test_tran/test_yell_agg_sheets2pidgin_group_raw_sheets.py
@@ -1,7 +1,7 @@
 from src.a00_data_toolbox.file_toolbox import create_path
 from src.a02_finance_logic._utils.strs_a02 import owner_name_str, fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, acct_name_str, event_int_str
-from src.a16_pidgin_logic.pidgin_config import (
+from src.a16_pidgin_logic._utils.str_a16 import (
     inx_bridge_str,
     otx_bridge_str,
     inx_label_str,
