+++ src/a19_world_logic/test_03_pidgin/test_world_etl03_04_otz_event_pidgin_df_to_pidgin_csvs.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import create_path
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
-from src.a16_pidgin_logic.pidgin_config import (
+from src.a16_pidgin_logic._utils.str_a16 import (
     inx_bridge_str,
     otx_bridge_str,
     inx_road_str,
