+++ src/a19_world_logic/test_03_pidgin/test_world_etl03_05_otz_event_pidgin_csvs_to_pidgin_json.py
@@ -7,10 +7,8 @@ from src.a00_data_toolbox.file_toolbox import (
 from src.a01_word_logic.road import default_bridge_if_None
 from src.a06_bud_logic._utils.str_a06 import type_RoadUnit_str
 from src.a16_pidgin_logic.pidgin import get_pidginunit_from_json
-from src.a16_pidgin_logic.pidgin_config import (
-    pidgin_filename,
-    default_unknown_word_if_None,
-)
+from src.a16_pidgin_logic._utils.str_a16 import pidgin_filename
+from src.a16_pidgin_logic.pidgin_config import default_unknown_word_if_None
 from src.a19_world_logic.world import worldunit_shop
 from src.a19_world_logic._utils.env_a19 import (
     env_dir_setup_cleanup,
