+++ src/a19_world_logic/test_04_ideas_pidgined/test_world_etl04_03_otz_event_ideas_to_otz_event_inx_ideas.py
@@ -7,7 +7,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     type_NameUnit_str,
     type_TagUnit_str,
 )
-from src.a16_pidgin_logic.pidgin_config import pidgin_filename
+from src.a16_pidgin_logic._utils.str_a16 import pidgin_filename
 from src.a16_pidgin_logic.pidgin import pidginunit_shop
 from src.a17_idea_logic.idea_db_tool import (
     upsert_sheet,
